% Generated by roxygen2 (4.0.2): do not edit by hand
\name{accuracy.glm.cross}
\alias{accuracy.glm.cross}
\title{Accuracy function to evaluate a gml model for training and test dataset}
\usage{
accuracy.glm.cross(m, abundance, test, depvar_name, abundance_name)
}
\arguments{
\item{m}{glm model}

\item{abundance}{abundances, should be passed in same order as presences and same length}

\item{test}{dataframe holding test data}

\item{depvar_name}{name of te dependent variable in the test dataframe}

\item{abundance_name}{name of te abundance variable in the test dataframe}
}
\value{
a vector of named arguments (n=number data, np=numer of presences, )
}
\description{
Evaluate according to multiple criteria a gml model for training and test dataset
}

