% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utils.r
\name{accuracy_me_simple}
\alias{accuracy_me_simple}
\title{Accuracy function to evaluate a maxent model}
\usage{
accuracy_me_simple(me, p, a, abundance = NULL)
}
\arguments{
\item{me}{maxent model}

\item{p}{presences}

\item{a}{absences}

\item{abundance}{abundances, should be passed in same order as presences and same length}
}
\value{
a vector of named arguments (n=number data, np=numer of presences, )
}
\description{
Evaluate according to multiple criteria a maxent model (\code{\link{maxent}})
}
