% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{calcArea}
\alias{calcArea}
\title{Area under a curve}
\usage{
calcArea(x, y)
}
\arguments{
\item{x}{vector holding the x values}

\item{y}{vector holding the y values}
}
\value{
area under the curve
}
\description{
Calculates the trapezoid area (boxes+traingles) under the curve y=f(x)
}
\seealso{
\code{\link{calcAreaLim}}
}
