% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cor2df.fire}
\alias{cor2df.fire}
\title{Correlations above a threshold, showing aicc's of a logistic according to fire presence}
\usage{
cor2df.fire(data, vars = NULL, fire, threshold, use = "everything",
  method = c("pearson", "kendall", "spearman"))
}
\arguments{
\item{data}{dataframe with the data}

\item{vars}{vector of column names or column numbers holding the variables to analyse. If not specified all the columns will be used.}

\item{fire}{column name of number holding fire presence [0/1]}

\item{threshold}{correlation threshold}

\item{use}{an optional character string giving a method for computing covariances in the
presence of missing values. This must be (an abbreviation of) one of the strings
"everything", "all.obs", "complete.obs", "na.or.complete", or "pairwise.complete.obs"}

\item{method}{a character string indicating which correlation coefficient
is to be computed. One of "pearson" (default), "kendall", or "spearman", can be abbreviated}
}
\value{
a list with \itemize{
\item cors: dataframe holding the variable pairs with a correlation higher than the specified threshold, and the relative aicc's'
\item aiccs: dataframe holding  allthe aicc's' for all variables
}
}
\description{
This function returns a dataframe with the variable pairs above a given correlation threshold, and the aic
value of a logistic model with fire occurrence
}
\details{
It is based on the \code{\link{cor}} function, but instead of a correlation matrix it returns
a dataframe with the pairwise combinations above a threshold.
}
\seealso{
\code{\link{cordf}}
}

