\name{dep.vars}
\alias{dep.vars}
\title{Dependent variable}
\usage{
dep.vars(formula)
}
\arguments{
  \item{formula}{formula to inspect, either as formula
  object or string}
}
\value{
name of the dependent variable
}
\description{
Extract the name of the dependent variable from formula
}
\seealso{
\code{\link{all.vars}} from base package to get all
variables
}

