% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{dep_vars}
\alias{dep_vars}
\title{Dependent variable}
\usage{
dep_vars(formula)
}
\arguments{
\item{formula}{formula to inspect, either as formula object or string}
}
\value{
name of the dependent variable
}
\description{
Extract the name of the dependent variable from formula
}
\seealso{
\code{\link{all.vars}} from base package to get all variables
}
