\name{dependent}
\alias{dependent}
\title{Dependent variable}
\usage{
  dependent(formula)
}
\arguments{
  \item{formula}{formula to inspect, either as formula
  object or string}
}
\value{
  name of the dependent variable
}
\description{
  Extract the name of the dependent variable from formula
}

