% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{filename}
\alias{filename}
\title{Filename without extension}
\usage{
filename(file)
}
\arguments{
\item{file}{name of the file}
}
\value{
file name without extension
}
\description{
Strips the extension form the filename.
}
