% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{fill.na}
\alias{fill.na}
\title{Fill gaps in a vector}
\usage{
\method{fill.na}{Vector}(x, method = c("linearize", "previous", "next"),
  maxgap = 2)
}
\arguments{
\item{x}{numeric vector}

\item{method}{how to fill in the gaps (default by linearization, otherwise by 
previous/next value duplication)}

\item{maxgap}{maximum number of consecutive NAs to fill. Any longer gaps will be left unchanged.}
}
\value{
numeric vector with filled 1-value gaps
}
\description{
Fill gaps with linearization (mean of the adjacent values) or
repetition of previous/next value, by using the \code{\link{na.approx}}
function.
}
