% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{getArgs}
\alias{getArgs}
\title{Extract and Load command line arguments into session}
\usage{
getArgs()
}
\value{
Nothing
}
\description{
Loads the command line arguments supplied when this R session was invoked
into the session environment.
}
