\name{getMElambdas}
\alias{getMElambdas}
\title{Extract lambda file values}
\usage{
  getMElambdas(m)
}
\arguments{
  \item{m}{maxent model}
}
\value{
  dataframe with the lambda values (what, lambda, min, max)
}
\description{
  Filla a dataframe with the lambda file values of a maxent
  model (\code{\link{dismo::maxent}})
}

