\name{getMEparNum}
\alias{getMEparNum}
\title{Count the number of parameters (with lambda!=0)}
\usage{
  getMEparNum(m)
}
\arguments{
  \item{m}{maxent model}
}
\value{
  number of parameters
}
\description{
  Cunts the number of parameters (with lambda!=0) of a
  maxent model (\code{\link{dismo::maxent}})
}

