\name{ind.vars}
\alias{ind.vars}
\title{Independent variable(s)}
\usage{
ind.vars(formula)
}
\arguments{
  \item{formula}{formula to inspect, either as formula
  object or string}
}
\value{
name of the independent variable(s)
}
\description{
Extract the name of the independent variable(s) from
formula
}
\seealso{
\code{\link{all.vars}} from base package to get all
variables
}

