\name{independents}
\alias{independents}
\title{Independent variable(s)}
\usage{
  independents(formula)
}
\arguments{
  \item{formula}{formula to inspect, either as formula
  object or string}
}
\value{
  name of the independent variable(s)
}
\description{
  Extract the name of the independent variable(s) from
  formula
}

