% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utils.r
\name{maxent.formula}
\alias{maxent.formula}
\title{Maxent with formula}
\usage{
maxent.formula(formula, data, ...)
}
\arguments{
\item{formula}{formula with dependent and independent variables}

\item{data}{dataframe holding the data}

\item{...}{additional parameters to pass to maxent}
}
\value{
Maxent model
}
\description{
Wrapper for Maxent to use a formula instead of data and presences
}
\seealso{
\code{\link{maxent}}
}
