\name{me.constants}
\alias{me.constants}
\title{Constants of a maxent model}
\usage{
  me.constants(m)
}
\arguments{
  \item{m}{either a maxent model or a character vector
  containing the lines of a maxent lambda file}
}
\value{
  dataframe holding the constants (what, value)
}
\description{
  Extract the constants of a maxent model
  (\code{\link{maxent}})
}

