\name{me.predict}
\alias{me.predict}
\title{Predict new values of a maxent model}
\usage{
  me.predict(m, data)
}
\arguments{
  \item{m}{maxent model}

  \item{data}{frame holding the data}
}
\value{
  predictions
}
\description{
  Calculates new values of a maxent model
  (\code{\link{maxent}})
}

