% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utils.r
\name{me_constants}
\alias{me_constants}
\title{Constants of a maxent model}
\usage{
me_constants(m)
}
\arguments{
\item{m}{either a maxent model or a character vector containing the lines of a maxent lambda file}
}
\value{
dataframe holding the constants (what, value)
}
\description{
Extract the constants of a maxent model (\code{\link{maxent}})
}
