% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utils.r
\name{me_lambdas}
\alias{me_lambdas}
\title{Extract lambda file values}
\usage{
me_lambdas(m)
}
\arguments{
\item{m}{either a maxent model or a character vector containing the lines of a maxent lambda file}
}
\value{
dataframe with the lambda values (what, lambda, min, max)
}
\description{
Fill a data frame with the lambda file values of a maxent model (\code{\link{maxent}})
}
