% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utils.r
\name{me_predict}
\alias{me_predict}
\title{Predict new values of a maxent model}
\usage{
me_predict(m, data)
}
\arguments{
\item{m}{either a maxent model or a character vector containing the lines of a maxent lambda file}

\item{data}{data frame holding the data}
}
\value{
predictions
}
\description{
Calculates new values of a maxent model (\code{\link{maxent}})
}
