% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utils.r
\name{models}
\alias{models}
\title{Models from formulae}
\usage{
models(prefix, formulae, suffix, envir = parent.frame(1))
}
\arguments{
\item{prefix}{string representation of the model prefix (e.g.  "glm(")}

\item{formulae}{list of formulae as string expressions}

\item{suffix}{string representation of the model suffix (e.g.  ", family="binomial")"   )}

\item{envir}{environment in which to evaluate the model expression. By default evaluates in the 
environment that calls this function}
}
\value{
a list of models
}
\description{
Build a list of models according to different formulae, a prefix and a suffix
}
