% Generated by roxygen2 (4.0.1): do not edit by hand
\name{replace.na}
\alias{replace.na}
\title{Replace NA's}
\usage{
replace.na(data, value)
}
\arguments{
\item{data}{vector, dataframe, list or matrix}

\item{value}{replacement value}
}
\value{
a data structure with the given value instead of NA's
}
\description{
Substitutes NA values in the given vector, dataframe, matrix or list
}

