% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{replace_na}
\alias{replace_na}
\title{Replace NA's}
\usage{
replace_na(data, value)
}
\arguments{
\item{data}{vector, dataframe, list or matrix}

\item{value}{replacement value}
}
\value{
a data structure with the given value instead of NA's
}
\description{
Substitutes NA values in the given vector, dataframe, matrix or list
}
