% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{rescale01}
\alias{rescale01}
\title{Rescale a vector of numbers between 0 and 1}
\usage{
rescale01(x, na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector to rescale}

\item{na.rm}{logical indicating whether missing values should be removed}
}
\value{
numeric vector with rescaled values
}
\description{
This function rescales the values of a numeric vector
between 0 and 1
}
