% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utils.r
\name{tpr}
\alias{tpr}
\title{True positive rate (sensitivity)}
\usage{
tpr(d, thr, depvar, occurrence = presence)
}
\arguments{
\item{d}{dataframe}

\item{thr}{threshold value}

\item{depvar}{column holding the model output}

\item{occurrence}{column holding presence [0/1]}
}
\value{
true positive rate (sensitivity)
}
\description{
True positive rate (sensitivity)
}
