% Generated by roxygen2 (4.0.1): do not edit by hand
\name{vpd}
\alias{vpd}
\title{Vapour pressure deficit}
\usage{
vpd(T, H)
}
\arguments{
\item{T}{numeric vector with air temperature values [C]}

\item{H}{numeric vector with air humidity values [\%]}
}
\value{
vapur pressure deficit
}
\description{
Calculate vapour pressure deficit from temperature and humidity
}

