% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.r
\name{yearplots}
\alias{yearplots}
\title{Multiple plots of daily data over years}
\usage{
<<<<<<< HEAD
yearplots(data, vars, year = year, doy = doy, what = c("data",
  "mean", "na"), rows = floor(length(vars)/cols + 1), cols = 4)
=======
<<<<<<< HEAD
yearplots(data, vars, year = year, doy = doy, what = c("data",
  "mean", "na"), rows = floor(length(vars)/cols + 1), cols = 4)
=======
yearplots(data, vars, year = year, doy = doy, what = c("data", "mean",
  "na"), rows = floor(length(vars)/cols + 1), cols = 4)
>>>>>>> 82ceb4ed070b06c8cdd8b89c7a7274ce80e84e99
>>>>>>> 756163c509f040cb5e87fc021d9af7695adab655
}
\arguments{
\item{data}{dataframe with data}

\item{vars}{character vector with variable (columns) names holding the data to summarise}

\item{year}{name of the column holding the year}

\item{doy}{name of the column holding the doy  [1-366]}

\item{what}{what to plot: all values, means over the years or presence of NA's}

\item{rows}{number of rows for the multiple plots. Defaults accordind to the number of}

\item{cols}{number of columns for the multiple plots. Defaults to 4}
}
\value{
plot
}
\description{
Plots summaries of daily data in a yearly plot
}
