% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utils.r
\name{accuracy_me_cross}
\alias{accuracy_me_cross}
\title{Accuracy function to evaluate a maxent model for training and test dataset}
\usage{
accuracy_me_cross(me, abundance, test = NULL, depvar_name, abundance_name)
}
\arguments{
\item{me}{maxent model}

\item{abundance}{abundances, should be passed in same order as presences and same length}

\item{test}{dataframe holding test data}

\item{depvar_name}{name of te dependent variable in the test dataframe}

\item{abundance_name}{name of te abundance variable in the test dataframe}
}
\value{
a vector of named arguments (n=number data, np=numer of presences, )
}
\description{
Evaluate according to multiple criteria a maxent model (\code{\link{maxent}}) for training and test dataset
}
