% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utils.r
\name{bkr}
\alias{bkr}
\title{Background proportion}
\usage{
bkr(d, thr, depvar)
}
\arguments{
\item{d}{dataframe}

\item{thr}{threshold value}

\item{depvar}{column holding the model output}
}
\value{
background proportion
}
\description{
Background proportion
}
