% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fill.1.na}
\alias{fill.1.na}
\title{Fill 1-value gaps in a vector}
\usage{
fill.1.na(x, method = c("linearize", "previous", "next"))
}
\arguments{
\item{x}{numeric vector}

\item{method}{how to fill in the gaps (default by linearization, otherwise by
previous/next value duplication)}
}
\value{
numeric vector with filled 1-value gaps
}
\description{
Fill gaps of single values with linearization (mean of the adjacent values) or
repetition of previous/next value.
}

