% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getArgs}
\alias{getArgs}
\title{Extract and Load command line arguments into session}
\usage{
getArgs()
}
\value{
Nothing
}
\description{
Loads the command line arguments supplied when this R session was invoked
into the session environment.
}

