% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ind.vars}
\alias{ind.vars}
\title{Independent variable(s)}
\usage{
ind.vars(formula, simplify = F)
}
\arguments{
\item{formula}{formula to inspect, either as formula object or string}

\item{simplify}{if terms such as I(x^2) should be retained as variable or not}
}
\value{
name of the independent variable(s)
}
\description{
Extract the name of the independent variable(s) from formula
}
\seealso{
\code{\link{all.vars}} from base package to get all variables
}

