\name{tpr}
\alias{tpr}
\title{True positive rate (sensitivity)}
\usage{
  tpr(d, thr, depvar_name = "y",
    occurrence_colname = "presence")
}
\arguments{
  \item{d}{dataframe}

  \item{thr}{threshold value}

  \item{depvar_name}{name of the column holding the model
  output}

  \item{occurrence_colname}{name of the column holding
  presence [0/1]}
}
\value{
  true positive rate (sensitivity)
}
\description{
  True positive rate (sensitivity)
}

